0. Print list

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -1,4 +1,5 @@
 #include "lists.h"
+#include <stdio.h>
 
 /**
  * print_listint - function that prints all elements in linked list
@@ -9,7 +10,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	size_t nbr = 0;
+	int nbr = 0;
 
 	if (h != Null)
 	{
